Split Add Environment into type selector + type-specific sub-pages
- AddEnvironment.tsx rewritten as TypeSelector page (Docker Socket / SSH Tunnel)
- New AddDockerSocketEnvironment.tsx with socket path form
- New AddSshTunnelEnvironment.tsx with SSH connection form
- Routes: /environments/add → type selector, /add/docker-socket, /add/ssh-tunnel
- Pattern: consistent with AddStackSourceSelect (TypeSelector component)
